Move flow controller classes to message package
