+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/contractimpl/DefaultHttpClientConnector.java
@@ -52,6 +52,7 @@ import org.wso2.transport.http.netty.contractimpl.sender.http2.Http2ConnectionMa
 import org.wso2.transport.http.netty.contractimpl.sender.http2.OutboundMsgHolder;
 import org.wso2.transport.http.netty.contractimpl.sender.http2.RequestWriteStarter;
 import org.wso2.transport.http.netty.contractimpl.sender.http2.TimeoutHandler;
+import org.wso2.transport.http.netty.message.ClientRemoteFlowControlListener;
 import org.wso2.transport.http.netty.message.Http2PushPromise;
 import org.wso2.transport.http.netty.message.Http2Reset;
 import org.wso2.transport.http.netty.message.HttpCarbonMessage;
